+++ civis/resources/__init__.py
@@ -1,3 +1,3 @@
-from ._resources import generate_classes
+from ._resources import generate_classes, get_api_spec
 
-__all__ = ["generate_classes"]
+__all__ = ["generate_classes", "get_api_spec"]